Update primitives ver 0.1.23
