+++ samples/peoplenet_detector/demo_performance.yml
@@ -30,16 +30,6 @@ parameters:
     # drawfunc's class init keyword arguments
     # will be available as drawfunc object's attributes
     kwargs:
-      # RGBA color value for bboxes of persons with face associated
-      person_with_face_bbox_color: [0, 255, 0, 255]
-      # RGBA color value for bboxes of persons with no face associated
-      person_no_face_bbox_color: [0, 0, 255, 255]
-      # RGBA color value for person label's background
-      person_label_bg_color: [216, 230, 255, 255]
-      # RGBA color value for person label's font
-      person_label_font_color:  [0, 0, 0, 255]
-      # width of person's bboxes border in pixels
-      bbox_border_width: 3
       # height of overlay dashboard in pixels
       overlay_height: 180
       # height of logo image in pixels
@@ -50,6 +40,35 @@ parameters:
       counters_height: 85
       # thickness of counter's font
       counters_font_thickness: 5
+      rendered_objects:
+        # Detector's element name from pipeline definition below
+        peoplenet:
+          # Detector's output object's label from pipeline definition
+          # Note that in this sample the analytics pyfunc assigns draw labels to objects
+          # If assigned, draw labels are used in place of the original object labels
+          # in the default draw func implementation
+          person_face:
+            bbox:
+              # RGBA color value for bboxes of persons with face associated
+              border_color: 00ff00ff
+              # width of person's bboxes border in pixels
+              thickness: 3
+            label:
+              # RGBA color values for person label's font and background
+              font_color: 000000ff
+              background_color: d8e6ffd9
+          person_noface:
+            bbox:
+              # RGBA color value for bboxes of persons with no face associated
+              border_color: 0000ffff
+              # width of person's bboxes border in pixels
+              thickness: 3
+            label:
+              # RGBA color values for person label's font and background
+              font_color: 000000ff
+              background_color: d8e6ffd9
+          face:
+            blur: true
 
 # pipeline definition
 pipeline:
