+++ savant/deepstream/drawfunc.py
@@ -1,25 +1,21 @@
 """Default implementation PyFunc for drawing on frame."""
-from typing import Any, Dict, Optional, Tuple
-import copy
+from typing import Any, Dict, Optional
 import pyds
 import cv2
-from savant_rs.primitives import (
-    BoundingBoxDraw,
-    ColorDraw,
-    LabelDraw,
-    DotDraw,
-    PaddingDraw,
+from savant_rs.draw_spec import (
     ObjectDraw,
-    LabelPosition,
+    LabelDraw,
+    BoundingBoxDraw,
+    DotDraw,
     LabelPositionKind,
 )
-
 from savant.meta.object import ObjectMeta
 from savant.deepstream.base_drawfunc import BaseNvDsDrawFunc
 from savant.deepstream.meta.frame import NvDsFrameMeta
 from savant.meta.bbox import RBBox
 from savant.meta.constants import UNTRACKED_OBJECT_ID
 from savant.utils.artist import Position, Artist
+from savant.utils.draw_spec import get_obj_draw_spec, get_default_draw_spec
 from savant.gstreamer import Gst  # noqa: F401
 from savant.deepstream.opencv_utils import nvds_to_gpu_mat
 
@@ -42,30 +38,8 @@ class NvDsDrawFunc(BaseNvDsDrawFunc):
                 for obj, obj_draw_spec_cfg in objects.items():
                     self.draw_spec[(unit, obj)] = get_obj_draw_spec(obj_draw_spec_cfg)
         else:
-            default_bbox_spec = BoundingBoxDraw(
-                border_color=ColorDraw(red=0, green=255, blue=0, alpha=255),
-            )
-            default_bg_color = ColorDraw(red=0, green=0, blue=0, alpha=255)
-            default_font_color = ColorDraw(red=255, green=255, blue=255, alpha=255)
-            default_font_scale = 0.5
-            self.default_spec_track_id = ObjectDraw(
-                bounding_box=default_bbox_spec,
-                label=LabelDraw(
-                    font_color=default_font_color,
-                    font_scale=default_font_scale,
-                    background_color=default_bg_color,
-                    format=['{label} #{track_id}'],
-                ),
-            )
-            self.default_spec_no_track_id = ObjectDraw(
-                bounding_box=default_bbox_spec,
-                label=LabelDraw(
-                    font_color=default_font_color,
-                    font_scale=default_font_scale,
-                    background_color=default_bg_color,
-                    format=['{label}'],
-                ),
-            )
+            self.default_spec_track_id = get_default_draw_spec(track_id=True)
+            self.default_spec_no_track_id = get_default_draw_spec(track_id=False)
 
         self.frame_streams = []
 
@@ -108,8 +82,8 @@ class NvDsDrawFunc(BaseNvDsDrawFunc):
                 continue
 
             if len(self.draw_spec) > 0:
-                if (obj_meta.element_name, obj_meta.label) in self.draw_spec:
-                    spec = self.draw_spec[(obj_meta.element_name, obj_meta.label)]
+                if (obj_meta.element_name, obj_meta.draw_label) in self.draw_spec:
+                    spec = self.draw_spec[(obj_meta.element_name, obj_meta.draw_label)]
                     spec = self.override_draw_spec(obj_meta, spec.copy())
                 else:
                     continue
@@ -209,98 +183,3 @@ class NvDsDrawFunc(BaseNvDsDrawFunc):
 
     def _blur(self, obj_meta: ObjectMeta, artist: Artist):
         artist.blur(obj_meta.bbox)
-
-
-def convert_hex_to_rgba(hex_color: str) -> Tuple[int, int, int, int]:
-    """Convert hex color to RGBA.
-    Hex color string is expected to be exactly 8 characters long.
-
-    :param hex_color: Hex color string
-    :return: RGBA color tuple
-    """
-    return tuple(int(hex_color[i : i + 2], 16) for i in (0, 2, 4, 6))
-
-
-def get_obj_draw_spec(config: dict) -> ObjectDraw:
-    bbox_draw = None
-    if 'bbox' in config:
-        # default green borders
-        border_color = config['bbox'].get('border_color', '00FF00FF')
-        # default transparent background
-        background_color = config['bbox'].get('background_color', '00000000')
-        # default border thickness 2
-        thickness = config['bbox'].get('thickness', 2)
-        # default no padding
-
-        if 'padding' in config['bbox']:
-            padding_draw = PaddingDraw(**config['bbox']['padding'])
-        else:
-            padding_draw = PaddingDraw()
-        bbox_draw = BoundingBoxDraw(
-            border_color=ColorDraw(*convert_hex_to_rgba(border_color)),
-            background_color=ColorDraw(*convert_hex_to_rgba(background_color)),
-            padding=padding_draw,
-            thickness=thickness,
-        )
-
-    central_dot_draw = None
-    if 'central_dot' in config:
-        # default green color
-        color = config['central_dot'].get('color', '00FF00FF')
-        # default radius 5
-        radius = config['central_dot'].get('radius', 5)
-        central_dot_draw = DotDraw(
-            color=ColorDraw(*convert_hex_to_rgba(color)),
-            radius=radius,
-        )
-
-    label_draw = None
-    if 'label' in config:
-        # default white font color
-        font_color = config['label'].get('font_color', 'FFFFFFFF')
-        # default transparent border
-        border_color = config['label'].get('border_color', '00000000')
-        # default black background
-        background_color = config['label'].get('background_color', '000000FF')
-        # default font scale 0.5
-        font_scale = config['label'].get('font_scale', 0.5)
-        # default font thickness 1
-        thickness = config['label'].get('thickness', 1)
-        # default format: {label}
-        label_format = config['label'].get('format', ['{label}'])
-
-        # rely on rust for defaults for label position
-        if 'position' in config['label']:
-            label_pos_kwargs = copy.deepcopy(config['label']['position'])
-            if 'position' in label_pos_kwargs:
-                if label_pos_kwargs['position'] == 'Center':
-                    label_pos_kwargs['position'] = LabelPositionKind.Center
-                elif label_pos_kwargs['position'] == 'TopLeftOutside':
-                    label_pos_kwargs['position'] = LabelPositionKind.TopLeftOutside
-                elif label_pos_kwargs['position'] == 'TopLeftInside':
-                    label_pos_kwargs['position'] = LabelPositionKind.TopLeftInside
-                else:
-                    # invalid position kind
-                    label_pos_kwargs.pop('position')
-            label_position = LabelPosition(**label_pos_kwargs)
-        else:
-            label_position = LabelPosition()
-
-        label_draw = LabelDraw(
-            font_color=ColorDraw(*convert_hex_to_rgba(font_color)),
-            border_color=ColorDraw(*convert_hex_to_rgba(border_color)),
-            background_color=ColorDraw(*convert_hex_to_rgba(background_color)),
-            font_scale=font_scale,
-            thickness=thickness,
-            format=label_format,
-            position=label_position,
-        )
-
-    blur = config.get('blur', False)
-
-    return ObjectDraw(
-        bounding_box=bbox_draw,
-        label=label_draw,
-        central_dot=central_dot_draw,
-        blur=blur,
-    )
